chore(ios): Remove Fabric dependency on RCTConvert due to RCTColorSpace
